Worked around an apparent bug in ar.
Changing a comment in a gmock header after a test was built would cause
it to fail to link when it was rebuilt, with "offset beyond end of file"
link errors. 'ar rv' is supposed to intelligently replace things, but it
seemed to be screwing up somehow. Instead, just start over from scratch
each time.