Fixed a build problem on Linux.
The version of google-glog that is installed by apt-get as of yesterday
(0.3.0-1) contains a bug where LOG_IF(..., DFATAL) fails to compile, due
to a missing namespace on a call to LogMessage. Switched to FATAL
instead; this can be reverted when the widely used verison is higher. It
appears to be fixed in 0.3.1.