Added accessors for gjstest.Matcher descriptions.
Deprecated the plain string properties. This is part of a change arc
that will end in expectThat, expectCall, and the built-in matchers not
computing descriptions unless a failure occurs. The motivation is to
avoid calling gjstest.stringify when unnecessary, because this can be
expensive for very large objects.

Future changes will remove all uses of the deprecated string properties,
then remove them altogether.