Made everything work with V8 version 3.28.71.
In the process, switched away from trying to make weak references to
registered functions. This was a pain in the ass when tracking API
changes, and necessitated all sorts of hacks to encourage garbage
collection in testing. These hacks weren't principled -- finalizers
still aren't guaranteed to run; in particular they seem to not run when
disposing an isolate. So don't try to pretend that we can avoid leaks
in the embedding program with this mechanism.